standardize card search
