+++ crates/opsml_server/opsml_ui/opsml-theme.css
@@ -63,7 +63,7 @@
 	--color-secondary-200: oklch(90.03% 0.08 164.54deg);
 	--color-secondary-300: oklch(86.05% 0.1 160.39deg);
 	--color-secondary-400: oklch(82.37% 0.13 156.92deg);
-	--color-secondary-500: oklch(78.92% 0.15 154.02deg);
+	--color-secondary-500: oklch(78.955% 0.14997 153.915);
 	--color-secondary-600: oklch(71.83% 0.15 153.14deg);
 	--color-secondary-700: oklch(64.6% 0.14 152.1deg);
 	--color-secondary-800: oklch(56.97% 0.14 151.29deg);
