+++ crates/opsml_server/opsml_ui/src/lib/components/card/CardTableView.svelte
@@ -56,14 +56,22 @@
       </thead>
       <tbody>
       {#each registryPage.summaries as summary, i}
-        <tr class={`border-b-2 border-black hover:bg-primary-300 ${i % 2 === 0 ? 'bg-gray-100' : 'bg-white'}`}>
-          <td class="p-2 pl-8">{summary.space}</td>
-          <td class="p-2 text-center">{summary.name}</td>
-          <td class="p-2 text-center">{summary.updated_at}</td>
-          <td class="p-2 text-center">{summary.version}</td>
-          <td class="p-2 text-center">{summary.versions}</td>
+        <tr class={`border-b-2 border-black hover:bg-primary-300 ${i % 2 === 0 ? 'bg-white' : 'bg-white'}`}>
+          <td class="p-1 pl-8">{summary.space}</td>
+          <td class="p-1 text-center">{summary.name}</td>
+          <td class="p-1 text-center ">
+            <div class="badge bg-secondary-100 text-black items-center gap-1 px-2 py-1">
+              {summary.updated_at}
+            </div>
+          </td>
+          <td class="p-1 text-center">
+            <div class="badge bg-error-100 text-black items-center gap-1 px-2 py-1">
+              {summary.version}
+            </div>
+          </td>
+          <td class="p-1 text-center">{summary.versions}</td>
           <td class="p-2">
-            <button class="btn text-sm flex flex-row gap-1 bg-primary-500 shadow shadow-hover border-black border-2 rounded-lg" onclick={() => navigateToCardPage(registry, summary.space, summary.name, summary.version)}>
+            <button class="justify-self-center btn text-sm flex flex-row gap-1 bg-primary-500 shadow shadow-hover border-black border-2 rounded-lg" onclick={() => navigateToCardPage(registry, summary.space, summary.name, summary.version)}>
               <div class="text-black">Link</div>
             </button>
           </td>
